requirements/common.txt
Flask==1.0.2
PyYAML==3.13
Flask-Mail==0.9.1